#1518 Small renames - branch fixes
