+++ data-model/src/main/scala/za/co/absa/enceladus/model/MappingTable.scala
@@ -17,9 +17,7 @@ package za.co.absa.enceladus.model
 
 import java.time.ZonedDateTime
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.node.{ArrayNode, ObjectNode}
-import com.fasterxml.jackson.module.scala.DefaultScalaModule
+import com.fasterxml.jackson.databind.node.ArrayNode
 import za.co.absa.enceladus.model.versionedModel.VersionedModel
 import za.co.absa.enceladus.model.menas.audit._
 import za.co.absa.enceladus.model.menas.MenasReference
